Update to match the firefox version

+++ push-to-pootle.sh
@@ -12,6 +12,8 @@ manage_command="/var/www/sites/$instance/src/manage.py"
 manage_py_verbosity=2
 precommand=". /var/www/sites/mozilla/env/bin/activate;"
 
+option_project="--project=$project"
+
 if [ $# -lt 1 ]; then
 	echo "$(basename $0) [lang(s)]"
 	exit
@@ -28,7 +30,6 @@ do
 	option_langs="$option_langs --language=$lang"
 done
 
-option_project="--project=$project"
 
 sync_command="$precommand python $manage_command sync_stores --verbosity=${manage_py_verbosity} $option_project $option_langs"
 update_command="$precommand python $manage_command update_stores $option_project"
@@ -46,17 +47,32 @@ fi
 for lang in $langs
 do
 	rm -rf $local_copy/$lang
-	cat $phaselist | while read phase file
-	do
-		mkdir -p $local_copy/$lang/$phase/$(dirname $file)
-		cp -p $lang/$file $local_copy/$lang/$phase/$file
-	done
+	if [ -f "$phaselist" ]; then
+		cat $phaselist | while read phase file
+		do
+			if [ $lang == "pot" -o $lang == "templates" ]; then
+				file=${file}t
+			fi
+			mkdir -p $local_copy/$lang/$phase/$(dirname $file)
+			cp -p $lang/$file $local_copy/$lang/$phase/$file
+		done
+	else
+		if [ $lang == "pot" -o $lang == "templates" ]; then
+			find $lang -name "*.pot"
+		else
+			find $lang -name "*.po"
+		fi | while read file
+		do
+			mkdir -p $local_copy/$(dirname $file)
+			cp -p $file $local_copy/$file
+		done
+	fi
 done
 
 
 for lang in $langs
 do
 	# FIXME only sync if we copied up correctly, this way we catch permission errors quickly
-	rsync -az --no-g --chmod=Dg+s,ug+rw,o-rw,Fug+rw,o-rw --include="*.po" --delete $local_copy/$lang $user@$server:$pootle_dir/
+	rsync -az --no-g --chmod=Dg+s,ug+rw,o-rw,Fug+rw,o-rw --include="*.po" --exclude=pootle-terminology.po --exclude=.translation_index --delete $local_copy/$lang $user@$server:$pootle_dir/
 	ssh $user@$server "$update_command --language=$lang"
 done
